requirement.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
pygame==2.5.0
python-dotenv==0.21.1
requests==2.31.0
urllib3==2.0.4
